@@ -14,13 +14,12 @@
         }
     ],
     "require": {
-        "php": "~5.5|~7.0",
+        "php": "^7.4 || ^8.0",
         "ext-bcmath": "*"
     },
     "require-dev": {
-        "phpunit/phpunit": "~4.8",
-        "scrutinizer/ocular": "~1.1",
-        "squizlabs/php_codesniffer": "~2.6"
+        "phpunit/phpunit": "^9.5",
+        "squizlabs/php_codesniffer": "^3.5"
     },
     "autoload": {
         "psr-4": {
