@@ -20,20 +20,20 @@ class HourTest extends TestCase
      *
      * @dataProvider validHourProvider
      */
-    public function testInstantiateSuccess($value)
+    public function testInstantiateSuccess($value): void
     {
         $month = new Hour($value);
-        $this->assertInstanceOf('Gentle\Embeddable\Time\Hour', $month);
+        $this->assertInstanceOf(Hour::class, $month);
     }
 
     /**
      * @param string|int $value
      *
      * @dataProvider invalidHourProvider
-     * @expectedException \InvalidArgumentException
      */
-    public function testInstantiateError($value)
+    public function testInstantiateError($value): void
     {
+        $this->expectException(\InvalidArgumentException::class);
         new Hour($value);
     }
 
@@ -41,10 +41,10 @@ class HourTest extends TestCase
      * @param string|int $value
      *
      * @dataProvider invalidHourRangeProvider
-     * @expectedException \OutOfRangeException
      */
-    public function testMonthRangeError($value)
+    public function testMonthRangeError($value): void
     {
+        $this->expectException(\OutOfRangeException::class);
         new Hour($value);
     }
 
@@ -53,7 +53,7 @@ class HourTest extends TestCase
      *
      * @dataProvider validHourProvider
      */
-    public function testEqualityByValue($value)
+    public function testEqualityByValue($value): void
     {
         $hour1 = new Hour($value);
         $hour2 = new Hour($value);
@@ -65,7 +65,7 @@ class HourTest extends TestCase
     /**
      * @return array
      */
-    public function validHourProvider()
+    public function validHourProvider(): array
     {
         return [
             ['04'], [2], ['12'], ['8'], ['09'], ['00'], [0]
@@ -75,14 +75,14 @@ class HourTest extends TestCase
     /**
      * @return array
      */
-    public function invalidHourProvider()
+    public function invalidHourProvider(): array
     {
         return [
             ['aa4b'], ['not valid'], ['43.12'], ['14.1'], ['43,01'], [new \stdClass()], ['4321a'], [['3412']],
         ];
     }
 
-    public function invalidHourRangeProvider()
+    public function invalidHourRangeProvider(): array
     {
         return [
             [999], [312], ['034'], ['100'], [24], ['24']
